Type fix

+++ TypeScript/src/models/favorites.model.ts
@@ -3,6 +3,6 @@ import { DataTypes } from 'sequelize';
 module.exports = (sequelize) => {
     return sequelize.define('favorites', {
         item_id: DataTypes.STRING,
-        freq: DataTypes.INTEGER
+        freq: DataTypes.FLOAT
     })
 };
